requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-cloud-firestore==2.19.0
